refactor: :art: rename clustering for proposal 1 to proposal 1, move that folder to clustering folder and do the same with selectKBestChi2
