+++ clustering/proposal_1/calculate_avg_rating.py
