+++ clustering/proposal_1/calculate_distance_between_users_centroid.py
