+++ clustering/proposal_1/calculate_scaling.py
