+++ clustering/proposal_1/calculate_sum_F_D.py
