+++ clustering/proposal_1/create_centroid.py
