+++ clustering/proposal_1/create_centroids_list.py
