+++ clustering/proposal_1/discard_nearest_points.py
