+++ clustering/proposal_1/output/D.txt
