+++ clustering/proposal_1/output/D_.txt
