+++ clustering/proposal_1/output/F.txt
