+++ clustering/proposal_1/output/F_D.txt
