+++ clustering/proposal_1/output/M_nearest_points.txt
