+++ clustering/proposal_1/output/avg_ratings.txt
