+++ clustering/proposal_1/output/centroids.txt
