+++ clustering/proposal_1/output/max_D.txt
