+++ clustering/proposal_1/output/max_F_D.txt
