+++ clustering/proposal_1/output/new_centroid.txt
