+++ clustering/proposal_1/output/new_centroids.txt
