+++ clustering/proposal_1/update_centroids.py
