+++ clustering/proposal_2_chi2/calculate_avg_and_sum.py
