+++ clustering/proposal_2_chi2/calculate_observed_value.py
