+++ clustering/proposal_2_chi2/output/O.txt
