+++ clustering/proposal_2_chi2/output/avg_sum.txt
