+++ clustering/proposal_2_chi2/output/centroids.txt
