+++ clustering/proposal_2_chi2/output/chi2.txt
