+++ clustering/proposal_2_chi2/output/class_probability.txt
