--- selectKBest_chi2/__init__.py
@@ -1,1 +1,0 @@
-from .main_hadoop import run_clustering_chi2
